Replace ceil with ceilf

+++ cpp_cuda_c/src/spline_psf_gpu.cu
@@ -532,7 +532,7 @@ auto kernel_roi(spline *sp, float *rois, const int npx, const int npy, const flo
     z_delta = zc - z0;
 
     int n_threads = min(1024, npx * npy);  // max number of threads per block
-    int n_blocks = ceil(static_cast<float>(npx * npy) / static_cast<float>(n_threads));
+    int n_blocks = ceilf(static_cast<float>(npx * npy) / static_cast<float>(n_threads));
 
     fAt3Dj<<<n_blocks, n_threads>>>(sp, rois, r, npx, npy, x0, y0, z0, phot, x_delta, y_delta, z_delta);
 
@@ -565,7 +565,7 @@ auto kernel_derivative_roi(spline *sp, float *rois, float *drv_rois, const int n
     z_delta = zc - z0;
 
     int n_threads = min(1024, npx * npy);  // max number of threads per block
-    int n_blocks = ceil(static_cast<float>(npx * npy) / static_cast<float>(n_threads));
+    int n_blocks = ceilf(static_cast<float>(npx * npy) / static_cast<float>(n_threads));
 
     kernel_derivative<<<n_blocks, n_threads>>>(sp, rois, drv_rois, r, npx, npy, x0, y0, z0, phot, bg, x_delta, y_delta, z_delta, add_bg);
